+++ services/api/src/Domain/Interfaces/IKeyService.cs
@@ -4,9 +4,9 @@ namespace Domain.Interfaces;
 
 public interface IKeyService
 {
-    public string ToBase64(MessageType messageType, ushort seed, int messageLength, ReadOnlySpan<byte> key,
+    public string ToBase64(MessageType messageType, int seed, int messageLength, ReadOnlySpan<byte> key,
         ReadOnlySpan<byte> iV);
 
-    public bool TryParse(string base64Key, out MessageType messageType, out ushort seed, out int messageLength,
+    public bool TryParse(string base64Key, out MessageType messageType, out int seed, out int messageLength,
         out byte[] key, out byte[] iV);
 }
